requirements.txt
flask>=3.0
anthropic>=0.40
PyMuPDF>=1.24
python-docx>=1.1
gunicorn>=21.0
